Added main_bologna.py and module csv_cleaning.py
The  bologna.csv file has been added.

Added a module csv_cleaning.py that performs a cleaning of the main .csv file in order to change date format and convert duration from string to average number. A Holiday column has alse been added, it tells wether the day is a working or a non-working one. The visitors columns has been also modified in order to be compliant with the project.

Added the main_bologna.py.