Resolved the issue found in main.py
Resolved the issue: https://github.com/nataliababic/swdevel-lab-hfarm/issues/35#issue-2039349574

The control of the file existence was moved upon function3 since it is the first function that uses the file upon running. Now if the file is not present it creates it otherwise it reads it directly.

Other things changed:
- Changed csv_cleaning.py in order to handle duration with values 0 and None
- Now Duration in csv_cleaning is made all lowercase at the start of the function
- Added other tests for csv_cleaning
- Removed from main.py the unused code and the unused functions